refactor: ProfileController's to apply inheritance
The BaseProfileController class is implemented, to avoid duplicity of code through inheritance.
